+++ hazelcast/src/main/java/com/hazelcast/internal/serialization/impl/SerializationServiceV1.java
@@ -20,14 +20,6 @@ import com.hazelcast.core.HazelcastJsonValue;
 import com.hazelcast.internal.nio.BufferObjectDataInput;
 import com.hazelcast.internal.serialization.Data;
 import com.hazelcast.internal.serialization.DataType;
-import com.hazelcast.internal.serialization.PortableContext;
-import com.hazelcast.internal.serialization.impl.defaultserializers.ConstantSerializers;
-import com.hazelcast.internal.serialization.impl.defaultserializers.ConstantSerializers.BooleanSerializer;
-import com.hazelcast.internal.serialization.impl.defaultserializers.ConstantSerializers.ByteSerializer;
-import com.hazelcast.internal.serialization.impl.defaultserializers.ConstantSerializers.StringArraySerializer;
-import com.hazelcast.internal.serialization.impl.portable.PortableContextImpl;
-import com.hazelcast.internal.serialization.impl.portable.PortableHookLoader;
-import com.hazelcast.internal.serialization.impl.portable.PortableSerializer;
 import com.hazelcast.internal.serialization.impl.defaultserializers.ArrayBlockingQueueStreamSerializer;
 import com.hazelcast.internal.serialization.impl.defaultserializers.ArrayDequeStreamSerializer;
 import com.hazelcast.internal.serialization.impl.defaultserializers.ArrayListStreamSerializer;
@@ -35,6 +27,10 @@ import com.hazelcast.internal.serialization.impl.defaultserializers.ArrayStreamS
 import com.hazelcast.internal.serialization.impl.defaultserializers.ConcurrentHashMapStreamSerializer;
 import com.hazelcast.internal.serialization.impl.defaultserializers.ConcurrentSkipListMapStreamSerializer;
 import com.hazelcast.internal.serialization.impl.defaultserializers.ConcurrentSkipListSetStreamSerializer;
+import com.hazelcast.internal.serialization.impl.defaultserializers.ConstantSerializers;
+import com.hazelcast.internal.serialization.impl.defaultserializers.ConstantSerializers.BooleanSerializer;
+import com.hazelcast.internal.serialization.impl.defaultserializers.ConstantSerializers.ByteSerializer;
+import com.hazelcast.internal.serialization.impl.defaultserializers.ConstantSerializers.StringArraySerializer;
 import com.hazelcast.internal.serialization.impl.defaultserializers.CopyOnWriteArrayListStreamSerializer;
 import com.hazelcast.internal.serialization.impl.defaultserializers.CopyOnWriteArraySetStreamSerializer;
 import com.hazelcast.internal.serialization.impl.defaultserializers.DelayQueueStreamSerializer;
@@ -51,6 +47,10 @@ import com.hazelcast.internal.serialization.impl.defaultserializers.PriorityQueu
 import com.hazelcast.internal.serialization.impl.defaultserializers.SynchronousQueueStreamSerializer;
 import com.hazelcast.internal.serialization.impl.defaultserializers.TreeMapStreamSerializer;
 import com.hazelcast.internal.serialization.impl.defaultserializers.TreeSetStreamSerializer;
+import com.hazelcast.internal.serialization.impl.portable.PortableContext;
+import com.hazelcast.internal.serialization.impl.portable.PortableContextImpl;
+import com.hazelcast.internal.serialization.impl.portable.PortableHookLoader;
+import com.hazelcast.internal.serialization.impl.portable.PortableSerializer;
 import com.hazelcast.nio.ObjectDataInput;
 import com.hazelcast.nio.serialization.ClassDefinition;
 import com.hazelcast.nio.serialization.ClassNameFilter;
@@ -98,6 +98,10 @@ import java.util.concurrent.LinkedTransferQueue;
 import java.util.concurrent.PriorityBlockingQueue;
 import java.util.concurrent.SynchronousQueue;
 
+import static com.hazelcast.internal.serialization.impl.DataSerializableSerializer.EE_FLAG;
+import static com.hazelcast.internal.serialization.impl.DataSerializableSerializer.IDS_FLAG;
+import static com.hazelcast.internal.serialization.impl.DataSerializableSerializer.isFlagSet;
+import static com.hazelcast.internal.serialization.impl.SerializationUtil.createSerializerAdapter;
 import static com.hazelcast.internal.serialization.impl.defaultserializers.ConstantSerializers.BooleanArraySerializer;
 import static com.hazelcast.internal.serialization.impl.defaultserializers.ConstantSerializers.CharArraySerializer;
 import static com.hazelcast.internal.serialization.impl.defaultserializers.ConstantSerializers.CharSerializer;
@@ -115,16 +119,12 @@ import static com.hazelcast.internal.serialization.impl.defaultserializers.Const
 import static com.hazelcast.internal.serialization.impl.defaultserializers.ConstantSerializers.StringSerializer;
 import static com.hazelcast.internal.serialization.impl.defaultserializers.ConstantSerializers.TheByteArraySerializer;
 import static com.hazelcast.internal.serialization.impl.defaultserializers.ConstantSerializers.UuidSerializer;
-import static com.hazelcast.internal.serialization.impl.DataSerializableSerializer.EE_FLAG;
-import static com.hazelcast.internal.serialization.impl.DataSerializableSerializer.IDS_FLAG;
-import static com.hazelcast.internal.serialization.impl.DataSerializableSerializer.isFlagSet;
 import static com.hazelcast.internal.serialization.impl.defaultserializers.JavaDefaultSerializers.BigDecimalSerializer;
 import static com.hazelcast.internal.serialization.impl.defaultserializers.JavaDefaultSerializers.BigIntegerSerializer;
 import static com.hazelcast.internal.serialization.impl.defaultserializers.JavaDefaultSerializers.ClassSerializer;
 import static com.hazelcast.internal.serialization.impl.defaultserializers.JavaDefaultSerializers.DateSerializer;
 import static com.hazelcast.internal.serialization.impl.defaultserializers.JavaDefaultSerializers.HazelcastJsonValueSerializer;
 import static com.hazelcast.internal.serialization.impl.defaultserializers.JavaDefaultSerializers.JavaSerializer;
-import static com.hazelcast.internal.serialization.impl.SerializationUtil.createSerializerAdapter;
 import static com.hazelcast.internal.util.MapUtil.createHashMap;
 
 public class SerializationServiceV1 extends AbstractSerializationService {
@@ -142,7 +142,6 @@ public class SerializationServiceV1 extends AbstractSerializationService {
         for (ClassDefinition cd : loader.getDefinitions()) {
             portableContext.registerClassDefinition(cd);
         }
-
         dataSerializerAdapter = createSerializerAdapter(
                 new DataSerializableSerializer(builder.dataSerializableFactories, builder.getClassLoader()));
         portableSerializer = new PortableSerializer(portableContext, loader.getFactories());
