+++ hazelcast/src/main/java/com/hazelcast/internal/serialization/impl/portable/PortableHook.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hazelcast.internal.serialization;
+package com.hazelcast.internal.serialization.impl.portable;
 
 import com.hazelcast.nio.serialization.ClassDefinition;
 import com.hazelcast.nio.serialization.PortableFactory;
