+++ hazelcast/src/test/java/com/hazelcast/internal/serialization/impl/portable/PortableTest.java
@@ -20,7 +20,6 @@ import com.hazelcast.config.SerializationConfig;
 import com.hazelcast.config.SerializerConfig;
 import com.hazelcast.internal.serialization.Data;
 import com.hazelcast.internal.serialization.InternalSerializationService;
-import com.hazelcast.internal.serialization.PortableContext;
 import com.hazelcast.internal.serialization.SerializationService;
 import com.hazelcast.internal.serialization.impl.CustomSerializationTest;
 import com.hazelcast.internal.serialization.impl.DefaultSerializationServiceBuilder;
