+++ hazelcast/src/test/java/com/hazelcast/test/compatibility/SamplingSerializationService.java
@@ -19,7 +19,7 @@ package com.hazelcast.test.compatibility;
 import com.hazelcast.core.ManagedContext;
 import com.hazelcast.partition.PartitioningStrategy;
 import com.hazelcast.internal.serialization.InternalSerializationService;
-import com.hazelcast.internal.serialization.PortableContext;
+import com.hazelcast.internal.serialization.impl.portable.PortableContext;
 import com.hazelcast.internal.nio.BufferObjectDataInput;
 import com.hazelcast.internal.nio.BufferObjectDataOutput;
 import com.hazelcast.nio.ObjectDataInput;
